Revert "refactoring (#5)"
This reverts commit d20083acbd7cf47808dc467dc474cad4420a25bb.
