--- app/src/main/java/com/mobileedu02/smartschoolmanager/datatSource/SmartSchoolDao.kt
@@ -1,4 +1,0 @@
-package com.mobileedu02.smartschoolmanager.datatSource
-
-interface SmartSchoolDao {
-}
