--- app/src/main/java/com/mobileedu02/smartschoolmanager/datatSource/remote/FirebaseUtil.kt
@@ -1,2 +1,0 @@
-package com.mobileedu02.smartschoolmanager.datatSource.remote
-
