--- app/src/main/java/com/mobileedu02/smartschoolmanager/ui/fragments/auth/signin/SignInViewModel.kt
@@ -1,7 +1,0 @@
-package com.mobileedu02.smartschoolmanager.ui.fragments.auth.signin
-
-import androidx.lifecycle.ViewModel
-
-class SignInViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
