--- app/src/main/java/com/mobileedu02/smartschoolmanager/ui/fragments/auth/signout/SignOutViewModel.kt
@@ -1,7 +1,0 @@
-package com.mobileedu02.smartschoolmanager.ui.fragments.auth.signout
-
-import androidx.lifecycle.ViewModel
-
-class SignOutViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
