--- app/src/main/java/com/mobileedu02/smartschoolmanager/ui/fragments/splashscreen/SplashScreenFragment.kt
@@ -1,32 +1,0 @@
-package com.mobileedu02.smartschoolmanager.ui.fragments.splashscreen
-
-import androidx.lifecycle.ViewModelProviders
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.mobileedu02.smartschoolmanager.R
-
-class SplashScreenFragment : Fragment() {
-
-    companion object {
-        fun newInstance() = SplashScreenFragment()
-    }
-
-    private lateinit var viewModel: SplashScreenViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.splash_screen_fragment, container, false)
-    }
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProviders.of(this).get(SplashScreenViewModel::class.java)
-        // TODO: Use the ViewModel
-    }
-
-}
