--- app/src/main/java/com/mobileedu02/smartschoolmanager/ui/fragments/splashscreen/SplashScreenViewModel.kt
@@ -1,7 +1,0 @@
-package com.mobileedu02.smartschoolmanager.ui.fragments.splashscreen
-
-import androidx.lifecycle.ViewModel
-
-class SplashScreenViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
